# Conflicts:
#	src/member/model/service/MemberService.java

+++ src/product/model/dao/ProductDAO.java
@@ -54,10 +54,9 @@ public class ProductDAO {
 				p.setpName(rset.getString("PNAME"));
 				p.setpInfo(rset.getString("PINFO"));
 				p.setPrice(rset.getInt("PRICE"));
-				p.setDiscount(rset.getInt("DISCOUNT"));
+				p.setDiscount(rset.getDouble("DISCOUNT"));
 				p.setStock(rset.getInt("STOCK"));
 				p.setPhoto(rset.getString("PHOTO"));
-
 				list.add(p);
 			}
 			
@@ -86,7 +85,7 @@ public class ProductDAO {
 				p.setpName(rset.getString("PNAME"));
 				p.setpInfo(rset.getString("PINFO"));
 				p.setPrice(rset.getInt("PRICE"));
-				p.setDiscount(rset.getInt("DISCOUNT"));
+				p.setDiscount(rset.getDouble("DISCOUNT"));
 				p.setStock(rset.getInt("STOCK"));
 				p.setPhoto(rset.getString("PHOTO"));
 
@@ -127,7 +126,6 @@ public class ProductDAO {
 				w.setStock(rset.getInt("stock"));
 				w.setDiscount(rset.getDouble("discount"));
 				w.setPhoto(rset.getString("photo"));
-
 				list.add(w);
 			}
 
@@ -224,7 +222,7 @@ public class ProductDAO {
 				p.setpName(rset.getString("PNAME"));
 				p.setpInfo(rset.getString("PINFO"));
 				p.setPrice(rset.getInt("PRICE"));
-				p.setDiscount(rset.getInt("DISCOUNT"));
+				p.setDiscount(rset.getDouble("DISCOUNT"));
 				p.setStock(rset.getInt("STOCK"));
 				p.setPhoto(rset.getString("PHOTO"));
